+++ Sprite.cpp
@@ -22,8 +22,8 @@ void CSprite::SetSize( const D3DXVECTOR2 & Size )
 	vSize			= Size;
 	D3DSURFACE_DESC sd;
 	pTexture->GetLevelDesc(0, &sd);
-	vScaling		= D3DXVECTOR2( vSize.x/sd.Width, vSize.y/sd.Height ) * RES_X;
-	vRotationCenter = vSize * RES_X / 2;
+	vScaling		= D3DXVECTOR2( vSize.x/sd.Width, vSize.y/sd.Height ) * (float)RES_X;
+	vRotationCenter = vSize * (float)RES_X / 2;
 }
 
 
@@ -32,5 +32,5 @@ void CSprite::Render( LPD3DXSPRITE pSprite ) const
 	D3DXVECTOR2 Position = vPosition - vSize/2;
 	if (vScaling.x < 0) Position.x += vSize.x;
 	if (vScaling.y < 0) Position.y += vSize.y;
-	pSprite->Draw( pTexture, NULL, &vScaling, &vRotationCenter, fRotation, &(Position * RES_X), dwColor );
+	pSprite->Draw( pTexture, NULL, &vScaling, &vRotationCenter, fRotation, &(Position * (float)RES_X), dwColor );
 }
